+++ src/Router.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 import Main from './pages/Main/Main';
 import Signup from './pages/Signup/Signup';
 import Login from './pages/Login/Login';
