+++ src/pages/ProductDetail/ProductRec.scss
@@ -13,7 +13,6 @@
   }
 
   .rec-product-wrap {
-    width: 10%;
     margin-top: 100px;
   }
 }
