+++ src/pages/ProductList/ProductList.scss
@@ -1,19 +1,16 @@
 .product-page-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
   width: 70%;
   margin: 0px auto;
   padding: 80px 0px 200px;
-  display: flex;
-  flex-direction: column;
   overflow: visible;
-  justify-content: center;
-  align-items: center;
 
   .product-list {
-    margin-top: 100px;
     display: grid;
     grid-template-columns: auto auto auto auto;
     margin-top: 100px;
-    display: grid;
-    grid-template-columns: auto auto auto auto;
   }
 }
